+++ tests/Neo.Plugins.L2Batch.UnitTests/Usings.cs
@@ -1,0 +1,5 @@
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Neo;
+global using Neo.L2;
+global using Neo.L2.Telemetry;
+global using Neo.Plugins.L2;